backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
python-dotenv==1.0.1
SQLAlchemy==1.4.41
Flask-Login==0.6.3
Werkzeug==2.0.3
PyJWT==2.1.0
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
gunicorn==20.1.0 